% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_get}
\alias{api_get}
\title{Grab your mailchimp api key.}
\usage{
api_get(api = NULL)
}
\arguments{
\item{api}{Character. Your private api key. If api is `NULL`, the environment variable `Sys.getenv("mailchimp_api")` is used.}
}
\description{
A function to grab the mailchimp api key from your .Renviron.
}
\examples{
api_get()

}
