% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_set}
\alias{api_set}
\title{Set your mailchimp api key as an environment variable.}
\usage{
api_set(api)
}
\arguments{
\item{api}{Character. Your private api key.}
}
\description{
A function to set your mailchimp api key as an environment variable in your .Renviron
}
\examples{
api_set(api = "Xhui89090sjioH77aknovfdn")

}
