% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{apiending}
\alias{apiending}
\title{Splits the api key to get server for api call}
\usage{
apiending(apikey)
}
\arguments{
\item{apikey}{API key for mailchimp account}
}
\value{
the server for use in later api call
}
\description{
Is a helperfunction to extract the server that needs to be called in the api call. The server is the ending of the api key.
}
\examples{

**## Not run:**
apiending(apikey = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx-usxx")
## End(**Not run**)
}

