% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mailchimp.R
\name{fetch_mailchimp}
\alias{fetch_mailchimp}
\title{fetch_mailchimp
A function to fetch Mailchimp data from the windsor.ai API}
\usage{
fetch_mailchimp(
  api_key,
  date_from = NULL,
  date_to = NULL,
  fields = c("campaign", "clicks", "spend", "impressions", "date")
)
}
\arguments{
\item{api_key}{Your api key to access Windsor.ai API}

\item{date_from}{The date from which to start getting data in format YYYY-MM-DD}

\item{date_to}{The date until which to start getting data in format YYYY-MM-DD}

\item{fields}{he fields fetched from the API for a given connector
See https://windsor.ai/api-fields/ for details.}
}
\value{
A data frame with the desired data
}
\description{
fetch_mailchimp
A function to fetch Mailchimp data from the windsor.ai API
}
\examples{
\dontrun{
my_mailchimp_data <- fetch_mailchimp(api_key = "your api key",
date_from = "2022-10-01",
date_to = "2022-10-02",
fields = c("campaign", "clicks",
"spend", "impressions", "date"))
}
}
