% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_lists.R
\name{get_list}
\alias{get_list}
\title{Get information about a specific list}
\usage{
get_list(list_id = NULL, apikey = NULL, user = "anystring", ...)
}
\arguments{
\item{list_id}{The unique id for the list. Find the list id with \code{\link{get_lists}}}

\item{apikey}{API key for mailchimp account}

\item{user}{User to use in authentication. Can be any string you like. Defaults to "anystring".}

\item{...}{Query string parameters. See available parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/#}{API documentation}.}
}
\value{
Response body parameters. See response parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/#}{API documentation}.
}
\description{
Get information about a specific list in your MailChimp account. Results include list members who have signed up but haven’t confirmed their subscription yet and \href{http://kb.mailchimp.com/lists/managing-subscribers/view-unsubscribed-and-cleaned-addresses}{unsubscribed or cleaned}.
}
\examples{

**## Not run:**
mylist <- get_list(apikey = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx-usxx", list_id = "xxxxxxxxxx")
## End(**Not run**)

}

