% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{get_list_member_goals}
\alias{get_list_member_goals}
\title{Get the last 50 Goal events for a member on a specific list}
\usage{
get_list_member_goals(list_id = NULL, subscriber_hash = NULL,
  apikey = NULL, user = "anystring", ...)
}
\arguments{
\item{list_id}{The unique id for the list. Find the list id with \code{\link{get_lists}}.}

\item{subscriber_hash}{The MD5 hash of the lowercase version of the list member’s email address.}

\item{apikey}{API key for mailchimp account.}

\item{user}{User to use in authentication. Can be any string you like. Defaults to "anystring".}

\item{...}{Query string parameters. See available parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/members/goals/}{API documentation}.}
}
\value{
Response body parameters. See response parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/members/goals/}{API documentation}.
}
\description{
Get the last 50 Goal events for a member on a specific list
}
\examples{

**## Not run:**
mymember_goals <- get_list_member_goals(apikey = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx-usxx", list_id = "xxxxxxxxxx", subscriber_hash = "xxxxxxxxxx")
## End(**Not run**)

}

