% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{get_list_member_note}
\alias{get_list_member_note}
\title{Get recent notes for a specific list member}
\usage{
get_list_member_note(list_id = NULL, subscriber_hash = NULL,
  apikey = NULL, user = "anystring", ...)
}
\arguments{
\item{list_id}{The unique id for the list. Find the list id with \code{\link{get_lists}}.}

\item{subscriber_hash}{The MD5 hash of the lowercase version of the list member’s email address.}

\item{apikey}{API key for mailchimp account.}

\item{user}{User to use in authentication. Can be any string you like. Defaults to "anystring".}

\item{...}{Query string parameters. See available parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/members/notes/}{API documentation}.}

\item{note_id}{The id for the note.}
}
\value{
Response body parameters. See response parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/members/notes/}{API documentation}.
}
\description{
Get recent notes for a specific list member
}
\examples{

**## Not run:**
mymember_note <- get_list_member_note(apikey = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx-usxx", list_id = "xxxxxxxxxx", subscriber_hash = "xxxxxxxxxx",  note_id = "xxxxxxxxxx")
## End(**Not run**)

}

