% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{get_list_members}
\alias{get_list_members}
\title{Get information about members in a list}
\usage{
get_list_members(list_id = NULL, apikey = NULL, user = "anystring", ...)
}
\arguments{
\item{list_id}{The unique id for the list. Find the list id with \code{\link{get_lists}}}

\item{apikey}{API key for mailchimp account}

\item{user}{User to use in authentication. Can be any string you like. Defaults to "anystring".}

\item{...}{Query string parameters. See available parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/members/#}{API documentation}.}
}
\value{
Response body parameters. See response parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/members/#}{API documentation}.
}
\description{
Get information about members in a specific MailChimp list.
}
\examples{

**## Not run:**
mymembers <- get_list_members(apikey = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx-usxx", list_id = "xxxxxxxxxx")
## End(**Not run**)

}

