% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{get_list_segment_members}
\alias{get_list_segment_members}
\title{Get information about all members in a list segment}
\usage{
get_list_segment_members(list_id = NULL, segment_id = NULL, apikey = NULL,
  user = "anystring", ...)
}
\arguments{
\item{list_id}{The unique id for the list. Find the list id with \code{\link{get_lists}}.}

\item{segment_id}{The unique id for the segment.}

\item{apikey}{API key for mailchimp account}

\item{user}{User to use in authentication. Can be any string you like. Defaults to "anystring".}

\item{...}{Query string parameters. See available parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/segments/#}{API documentation}.}
}
\value{
Response body parameters. See response parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/segments/#}{API documentation}.
}
\description{
Get information about all members in a list segment
}
\examples{

**## Not run:**
mysegment_members <- get_list_segment_members(apikey = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx-usxx", list_id = "xxxxxxxxxx", segment_id = "xxxxxxxxxx")
## End(**Not run**)

}

