% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_lists.R
\name{get_list_segments}
\alias{get_list_segments}
\title{Get information about all segments in a list}
\usage{
get_list_segments(list_id = NULL, apikey = NULL, user = "anystring",
  content_type = "application/json", ...)
}
\arguments{
\item{list_id}{The unique id for the list. Find the list id with \code{\link{get_lists}}}

\item{apikey}{API key for mailchimp account}

\item{user}{User to use in authentication. Can be any string you like. Defaults to "anystring".}

\item{content_type}{MIME type (aka internet media type) used to override the content type returned by the server. See \url{http://en.wikipedia.org/wiki/Internet_media_type} for a list of common types.}

\item{...}{Query string parameters. See available parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/segments/#}{API documentation}.}
}
\value{
Response body parameters. See response parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/segments/#}{API documentation}.
}
\description{
Get information about all segments in a list
}
\examples{

**## Not run:**
mysegments <- get_list_segments(apikey = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx-usxx", list_id = "xxxxxxxxxx")
## End(**Not run**)

}
