% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_lists.R
\name{get_lists}
\alias{get_lists}
\title{Get information about all lists}
\usage{
get_lists(apikey = NULL, user = "anystring", type = "application/json",
  ...)
}
\arguments{
\item{apikey}{API key for mailchimp account}

\item{user}{User to use in authentication. Can be any string you like. Defaults to "anystring".}

\item{type}{MIME type (aka internet media type) used to override the content type returned by the server. See \url{http://en.wikipedia.org/wiki/Internet_media_type} for a list of common types.}

\item{...}{Query string parameters. See available parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/#}{API documentation}.}
}
\value{
Response body parameters. See response parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/#}{API documentation}.
}
\description{
Get information about all lists in the account.
}
\examples{

**## Not run:**
mylists <- get_lists(apikey = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx-usxx")
## End(**Not run**)

}
