% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_batch_sub}
\alias{list_batch_sub}
\title{Batch subscribe members to a list}
\usage{
list_batch_sub(api = NULL, list_id, new_subs)
}
\arguments{
\item{api}{Character. Your private api key. If api is `NULL`, the environment variable `Sys.getenv("mailchimp_api")` is used.}

\item{list_id}{Character. The ID of a list. See `get_lists()`.}

\item{new_subs}{Data Frame (or tibble). Email column must be named `email_address`.
Merge fields must match merge fields in `list_get_merge_fields()`}
}
\description{
Batch subscribe members to a list.
}
