% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_create}
\alias{list_create}
\title{Create a list/audience}
\usage{
list_create(
  api = NULL,
  name,
  company,
  address1,
  address2 = "",
  city,
  state,
  zip,
  country,
  phone = "",
  permission_reminder,
  from_name,
  from_email,
  subject = "",
  language = "en",
  email_type_option = FALSE
)
}
\arguments{
\item{api}{Character. Your private api key. If api is `NULL`, the environment variable `Sys.getenv("mailchimp_api")` is used.}

\item{name}{Name of the list.}

\item{company}{Character. Company name associated with the list.}

\item{address1}{Character. First line of the address of the associated company.}

\item{address2}{Character. Second line of the address of the associated company.}

\item{city}{Character. City of the associated company.}

\item{state}{Character. State of the associated company.}

\item{zip}{Character. Zip/post code of the associated company.}

\item{country}{Character Country of the associated company.}

\item{phone}{Character. Phone number of the associated company.}

\item{permission_reminder}{Character. Permission reminder for subscribers of the list.}

\item{from_name}{Character. Default from-name for emails to subscribers of the list.}

\item{from_email}{Character. Default from-email for emails to subscribers of the list.}

\item{subject}{Character. Default from-subject for emails to subscribers of the list.}

\item{language}{Character. Default language for emails to subscribers of the list.}

\item{email_type_option}{Boolean. See mailchimp api reference.}
}
\description{
Create a list/audience.
}
