% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_create_new_sub_list}
\alias{list_create_new_sub_list}
\title{Create list for one new list subscriber}
\usage{
list_create_new_sub_list(api = NULL, list_id, new_sub)
}
\arguments{
\item{api}{Character. Your private api key. If api is `NULL`, the environment variable `Sys.getenv("mailchimp_api")` is used.}

\item{list_id}{Character. The ID of a list. See `get_lists()`.}

\item{new_sub}{Data Frame (or tibble). Email column must be named `email_address`.
Merge fields must match merge fields in `list_get_merge_fields()`}
}
\value{
A list
}
\description{
Create list for one new list subscriber.
This is in prep for creating the JSON for multiple new subscribers in `list_batch_sub()`
}
