% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_get}
\alias{list_get}
\title{Grab a tibble of your mailchimp lists.}
\usage{
list_get(api = NULL)
}
\arguments{
\item{api}{Character. Your private api key. If api is `NULL`, the environment variable `Sys.getenv("mailchimp_api")` is used.}
}
\description{
Grab a tibble of your mailchimp lists.
}
