% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_get_members}
\alias{list_get_members}
\title{Returns information about the members of a particular list.}
\usage{
list_get_members(api = NULL, list_id, count = 10)
}
\arguments{
\item{api}{Character. Your private api key. If api is `NULL`, the environment variable `Sys.getenv("mailchimp_api")` is used.}

\item{list_id}{Character. The ID of a list. See `get_lists()`.}

\item{count}{Numeric. The number of members you want information about. Max 100 due to API rules.}
}
\description{
Returns information about the members of a particular list, such as EMAIL, merge fields and locations.
}
