% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_get_segments}
\alias{list_get_segments}
\title{Get the segments for a list}
\usage{
list_get_segments(api = NULL, list_id)
}
\arguments{
\item{api}{Character. Your private api key. If api is `NULL`, the environment variable `Sys.getenv("mailchimp_api")` is used.}

\item{list_id}{Character. The ID of a list. See `get_lists()`.}
}
\description{
Get the segments for a list
}
