% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-fields.R
\name{merge_field_create}
\alias{merge_field_create}
\title{Create a merge field for a list}
\usage{
merge_field_create(
  api = NULL,
  list_id,
  name,
  type = c("text", "number", "address", "phone", "date", "url", "image", "url", "radio",
    "dropdown", "birthday", "zip"),
  required = FALSE
)
}
\arguments{
\item{api}{Character. Your private api key. If api is `NULL`, the environment variable `Sys.getenv("mailchimp_api")` is used.}

\item{list_id}{Character. The ID of a list. See `list_get()`.}

\item{name}{Character. The name of the merge_field. See `merge_field_get()`.}

\item{type}{Character. The type of the merge field. Can be one of "text", "number", "address",
"phone", "date", "url", "image", "url", "radio", "dropdown", "birthday" or "zip".}

\item{required}{Boolean. Is this merge field required?}
}
\description{
Create a merge field for a list
}
