% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-fields.R
\name{merge_field_delete}
\alias{merge_field_delete}
\title{Delete a merge field from a list}
\usage{
merge_field_delete(api = NULL, list_id, merge_id)
}
\arguments{
\item{api}{Character. Your private api key. If api is `NULL`, the environment variable `Sys.getenv("mailchimp_api")` is used.}

\item{list_id}{Character. The ID of a list. See `list_get()`.}

\item{merge_id}{Character. The ID of a merge_field See `merge_field_get()`.}
}
\description{
Delete a merge field form a list
}
