% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-fields.R
\name{merge_field_get}
\alias{merge_field_get}
\title{Grab the merge fields of a list}
\usage{
merge_field_get(api = NULL, list_id)
}
\arguments{
\item{api}{Character. Your private api key. If api is `NULL`, the environment variable `Sys.getenv("mailchimp_api")` is used.}

\item{list_id}{Character. The ID of a list. See `get_lists()`.}
}
\description{
Grab the merge fields of a list.
}
