% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_lists.R
\name{post_segment}
\alias{post_segment}
\title{Create a new segment}
\usage{
post_segment(list_id = NULL, name = NULL, static_segment = NULL,
  apikey = NULL, user = "anystring", ...)
}
\arguments{
\item{name}{The name of the segment.}

\item{static_segment}{Vector of emails. Any emails provided that are not present will be ignored. Passing an empty vector will create a static segment without any subscribers. This field cannot be provided with the options field.}

\item{apikey}{API key for mailchimp account}

\item{user}{User to use in authentication. Can be any string you like. Defaults to "anystring".}

\item{...}{Query string parameters. See available parameters under CREATE here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/segments/}{API documentation}.}
}
\value{
Response body parameters. See response parameters under CREATE here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/segments/}{API documentation}.
}
\description{
Create a new segment in a specific list.
}
\examples{

**## Not run:**
post_segment <- post_segment(name = "my_new_segment", static_segment = c("e@mail1.com", "e@mail2.com"),  apikey = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx-usxx")
## End(**Not run**)

}

