% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_lists.R
\name{post_segment_members}
\alias{post_segment_members}
\title{Batch add/remove list members to static segment}
\usage{
post_segment_members(list_id = NULL, segment_id = NULL,
  members_to_add = NULL, members_to_remove = NULL, apikey = NULL,
  user = "anystring", ...)
}
\arguments{
\item{segment_id}{Id of segment.}

\item{members_to_add}{A vector of emails to be used for a static segment. Any emails provided that are not present on the list will be ignored.}

\item{members_to_remove}{A vector of emails to be used for a static segment. Any emails provided that are not present on the list will be ignored.}

\item{apikey}{API key for mailchimp account}

\item{user}{User to use in authentication. Can be any string you like. Defaults to "anystring".}

\item{...}{Query string parameters. See available parameters under CREATE here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/segments/}{API documentation}.}
}
\value{
Response body parameters. See response parameters under CREATE here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/segments/}{API documentation}.
}
\description{
Batch add/remove list members to static segment
}
\examples{

**## Not run:**
post_segment_members <- post_segment_members(list_id = "xxxxxxxxxx", members_to_add = c("e@mail1.com", "e@mail2.com"), members_to_remove = c("e@mail3.com", "e@mail4.com"), apikey = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx-usxx")
## End(**Not run**)

}

