% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{get_lists}
\alias{get_lists}
\title{Get information about all lists}
\usage{
get_lists(apikey = NULL, user = "anystring", ...)
}
\arguments{
\item{apikey}{API key for mailchimp account}

\item{user}{User to use in authentication. Can be any string you like. Defaults to "anystring".}

\item{...}{Query string parameters. See available parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/#}{API documentation}.}
}
\value{
Response body parameters. See response parameters under READ here: \href{https://developer.mailchimp.com/documentation/mailchimp/reference/lists/#}{API documentation}.
}
\description{
Get information about all lists in the account.
}
\examples{

**## Not run:**
mylists <- get_lists(apikey = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx-usxx")
## End(**Not run**)

}

